requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
jinja2==3.1.4
pydantic==2.9.2
apscheduler==3.10.4
python-dateutil==2.9.0
